🛂 feat: add support for oauth provider v0 (#150)
* :passport_control: feat: add list auth_providers

* :passport_control: feat: add list auth for google

* :hammer: feat: add `default_config.json` for postHook

* :hammer: feat: created `AuthProvider` registry

* :bug: fix: error on nextToken blank

* :wrench: nit: remove `createOauthUser` function

* :wrench: chore: comment in migration, move onboardUrl to config and self register auth provider

* :wrench: chore: add `applicationAuthenticationConfiguration` function to `Configuration.kt` from `Application.kt`

* :hammer: fix: lint issues

* :pencil2: fix: typo in config

* :pencil2: fix: dependency injection name and use store to use context for txns

+++ src/main/kotlin/com/hypto/iam/server/service/UsersService.kt
@@ -45,6 +45,7 @@ class UsersServiceImpl : KoinComponent, UsersService {
     private val appConfig: AppConfig by inject()
     private val passcodeRepo: PasscodeRepo by inject()
 
+    @Suppress("CyclomaticComplexMethod")
     override suspend fun createUser(
         organizationId: String,
         loginAccess: Boolean,
@@ -76,7 +77,7 @@ class UsersServiceImpl : KoinComponent, UsersService {
         }
 
         return txMan.wrap {
-            if (loginAccess) {
+            if (password != null && loginAccess) {
                 val passwordCredentials = PasswordCredentials(
                     username = username,
                     preferredUsername = preferredUsername,
@@ -120,8 +121,10 @@ class UsersServiceImpl : KoinComponent, UsersService {
                     policies.map { hrnFactory.getHrn(it) }
                 )
             }
-            email?.let {
-                passcodeRepo.deleteByEmailAndPurpose(email, VerifyEmailRequest.Purpose.invite)
+            if (password != null) {
+                email?.let {
+                    passcodeRepo.deleteByEmailAndPurpose(email, VerifyEmailRequest.Purpose.invite)
+                }
             }
             getUser(userHrn, userRecord)
         }
